training/requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
onnx>=1.15.0
onnxruntime>=1.17.0
optimum>=1.16.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
torchvision>=0.15.0
tqdm>=4.65.0
evaluate>=0.4.0
